Reworked signaling of presence
the old code would always signal presence after startup, which would lead
to errors when the power is cylced during ignition.

The changes will now wait 2s to see if a Init or InitIgn comes.
If not, presence is signalled every 2s, until:

a) we see Init
b) we see InitIgn and all devices where initialized

For (b), the return value of handleCmd is now bool.
A return value of true means that all devices are now initialized
(generated for both InitIgn and Init, but only interesting for InitIgn)

Of course now I need to differ between InitIgn and Init...